Initial commit
This code is being split out of the original shared repo we used for all
Ymir code. At this stage the application is a working backend written in
Flask. This application provides a restful API with some basic useful
resources for RPGs.

Currently, we are also hosting the API spec in the  /raml folder of this
repository.