Remove information from README not related to API
The README file was copied over from the master repository and still had
a lot of brainstorming about the project as a whole. This patch
eleminates everything that was not related to the backend. It also
updates the Travis CI badge to point to the badge for this repository.
In the future this README should be updated with usage instructions and
links to other parts of the project.